+++ internal/notify/notify_other.go
@@ -3,3 +3,4 @@
 package notify
 
 func Error(string, string) {}
+func Info(string, string)  {}